Better control of model rebuild
The model only needs to be rebuilt when the resolution of field types in an extensible model modifies the field type. This modification only occurs if the field is itself of an extensible type. In this case, the original type is replaced by the class aggregated for the original type.